Fix for memory leak
When local.pipe(remote) produces an error, events were not properly cleaned up on the remote stream. This resulted in a memory leak and eventually the client crashes. This bubbled up and produced the "connection refused: ... (check your firewall settings)" error. For example, by default after 2 minutes of an idle socket connection the socket produces a reconnect error. Every time this happens, events are leftover on the remote connection, leaking memory. The remote connection should only be closed if the local connection is refused, otherwise the local should unpipe remote and remote should try to reestablish connection with local.